Change size_t usage to uint16_t.
The uint16_t data type is more that enough to hold size and offset values
in HAN-FUN.

+++ apps/base.h
@@ -205,7 +205,7 @@ struct Base:public HF::Devices::Concentrator::Abstract <Unit0>
    virtual ~Base()
    {}
 
-   void receive (HF::Protocol::Packet &packet, HF::Common::ByteArray &payload, size_t offset);
+   void receive (HF::Protocol::Packet &packet, HF::Common::ByteArray &payload, uint16_t offset);
 
    using HF::Devices::Concentrator::Abstract <Unit0>::link;
