+++ inc/hanfun/core.h
@@ -338,7 +338,7 @@ namespace HF
       // =============================================================================
 
       //! @copydoc HF::Interface::handle
-      Common::Result handle (HF::Protocol::Packet &packet, Common::ByteArray &payload, size_t offset)
+      Common::Result handle (HF::Protocol::Packet &packet, Common::ByteArray &payload, uint16_t offset)
       {
          Core::IService *service = find <0, ITF...>(packet.message.itf.id);
